requirements.txt
Scrapy==0.22.2
python-dateutil==2.2